Expose entities in query as views
For security reasons, it's dangerous and not so useful to expose
every table as-is. Therefore, the tables should be provided as
views, which decides the list of items to show according to the
context(e.g. viewer).

There are two types of views currently, which are Direct View and
ID View. Direct View shows every content in the table and provides
basic filtering, ordering and pagination. It's useful for admin.
On the other hand, there's ID View, that developers decide which
IDs to return according to the context and arguments. Currently
this view only supports single primary column.